# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskFlow API",
        "description": "API REST de gestion de proyectos y tareas. Backend puro: sin vistas y sin\nsesiones, pensada para que la consuma cualquier cliente.\n\n**Autenticacion.** Registrate o inicia sesion para recibir un token, y\nmandalo en cada peticion como `Authorization: Bearer <token>`. Los tokens\ncaducan a los siete dias.\n\n**Aislamiento.** Cada usuario solo ve y edita sus propios proyectos. Pedir\nun recurso de otra persona devuelve 403; las tareas heredan el permiso del\nproyecto al que pertenecen.\n\n**Datos de prueba.** Tras `php artisan migrate:fresh --seed` existe el\nusuario `demo@taskflow.test` con la contrasena `contrasena-demo`, y un\ncuarto proyecto que pertenece a otra persona para poder comprobar el 403.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/projects/{project}": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Devuelve un proyecto con su numero de tareas",
                "operationId": "projects.show",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proyecto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sin token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "El proyecto existe pero es de otra persona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe, o fue borrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Borra un proyecto y arrastra sus tareas",
                "description": "Borrado logico. Las tareas del proyecto quedan borradas con el y dejan de responder.",
                "operationId": "projects.destroy",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Borrado"
                    },
                    "401": {
                        "description": "Sin token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Proyecto de otra persona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Actualiza un proyecto",
                "description": "Actualizacion parcial: los campos omitidos se quedan como estaban. PUT se acepta como alias.",
                "operationId": "projects.update",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 2000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sin token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Proyecto de otra persona. Se comprueba antes de validar, asi que llega este 403 y no un 422.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos invalidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "title": "Proyecto",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "01a045d6-b05e-70c6-a3c9-29454f3eb0ad"
                    },
                    "name": {
                        "type": "string",
                        "example": "Rediseno del sitio"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "tasks_count": {
                        "description": "Solo aparece cuando el endpoint lo calcula; su ausencia no significa cero.",
                        "type": "integer",
                        "example": 6
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "Error",
                "description": "Cuerpo de todo error salvo los de validacion.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Error de validacion",
                "description": "Unico cuerpo de error con estructura anidada.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token devuelto por /auth/register o /auth/login.",
                "scheme": "bearer"
            }
        }
    },
...
}